Revert xpath version and move the dependency to lib
The npm xpath package has been updated to more closely match
the xpath spec. However, this leads to problems using generic
xpath selectors to select into a DOM rendered from normal HTML.
Selectors that don't have a namespace fail to select anything.

This behaviour makes scraping of the kind enabled by thresher
impossible, but the previous version of the xpath lib was still
the best working implementation we found in npm during testing.

The fix is to bring the last working version of the xpath lib
into this repo (it's only a single file). At a later stage, we
can take later versions of the xpath lib that might have other
improvements, but restore the historical handling of namespaces.
This is tracked in issue #17.