Fix bug in recursive scope for follow-ons
The Scraper object was storing a single renderer as av ivar.

Inside scrapeURL, when resursively climbing down the element tree,
the 'result' event of the renderer was being written to each time.

This was leading to the event function from the highest level of
recursion being retained in scope at every deeper level, so that
child elements were always those of the highest level node.

The fix was to merely set the renderer type at the Scraper init,
then to create a new instance of the renderer at every round of
recursion, thereby creating a new scope with the expected node.